+++ src/app/app.component.ts
@@ -1,10 +1,16 @@
-import { Component } from '@angular/core';
+import { Component, AfterViewInit, ViewChild, ElementRef } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
-  title = 'Candy Shop';
+export class AppComponent implements AfterViewInit {
+
+  @ViewChild('appTitle')
+  title: ElementRef;
+
+  ngAfterViewInit() {
+    this.title.nativeElement.innerHTML = 'Candy Shop';
+  }
 }