+++ src/app/app.module.ts
@@ -6,11 +6,12 @@ import { AppComponent } from './app.component';
 import { ProductModule } from './product/product.module';
 import { CartModule } from './cart/cart.module';
 import { OrderModule } from './order/order.module';
+import { HighlightDirective } from './shared/highlight.directive';
+import { NO_ERRORS_SCHEMA } from '@angular/compiler/src/core';
+import { SharedModule } from './shared/shared.module';
 
 @NgModule({
-  declarations: [
-    AppComponent
-  ],
+  declarations: [AppComponent],
   imports: [
     BrowserModule,
     MatButtonModule,
@@ -18,10 +19,10 @@ import { OrderModule } from './order/order.module';
     MatTableModule,
     ProductModule,
     CartModule,
-    OrderModule
+    OrderModule,
+    SharedModule
   ],
-  providers: [],
-  bootstrap: [AppComponent],
+  bootstrap: [AppComponent]
 })
 export class AppModule {
 }