+++ src/app/cart/cart.module.ts
@@ -2,17 +2,20 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CartComponent } from './components/cart/cart.component';
 import { CartListComponent } from './components/cart-list/cart-list.component';
+import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
+    SharedModule
   ],
   declarations: [
     CartComponent,
     CartListComponent
   ],
   exports: [
-    CartComponent, CartListComponent
+    CartComponent, 
+    CartListComponent
   ]
 })
 export class CartModule {